Refactor: 디버깅 및 기능 개선
Event1Talk : OnTriggerEnter2D( )를 공통적으로 사용하고 있는 듯 하여 EventTalk로 스크립트명 변경
Coroutine 사용하여 gameObject가 비활성화 되기까지 약간의 딜레이 추가
Unity에서 각 Trigger 영역에 wall 추가하여 이벤트 시작했을 때 캐릭터가 해당 범위 안에 제대로 멈춰있을 수 있도록 조정

GameManager: 룬 관련 변수 삭제, 보스 정보에서 속성요소 삭제, 보스 클리어 여부 확인 변수 삭제

Hacking: 보스 클리어 여부를 보스 인덱스로 파악하게 변경되었으므로 이에 따른 코드 수정

EnterBossRoom: 보스 속성 요소가 삭제되었으므로 이에 따른 코드 수정

BossDogscene: 보스 인덱스로 파악할 수 있도록 수정 - 154 line

사용되지 않는 스크립트 정리

Tutorial 창이 Canvas에서 보이지 않는 현상이 있어 Image 오브젝트로 변환하고 부착
Event 관련된 스크립트들을 모두 EventManager라는 빈 오브젝트에 부착하여 관리
