+++ EPOH/Assets/Scripts/GameManager.cs
@@ -9,15 +9,13 @@ public class GameManager : MonoBehaviour
     public bool if_first = true; // Start 씬의 '시작 버튼'최초 클릭 여부 기록
 
     public int story_info = 0; // 스토리 진행 정도에 대한 정보
-    public bool[] rune = new bool[5]; // 룬이 활성화 되었는가에 대한 정보. 순서대로 Anger, Fear, Humiliation, Sorrow, Regret
+    public int tutorial_info = 0;
 
-    public int boss_num = -1; //위젯에서 선택한 보스의 인덱스
-    public const int boss_cnt = 3; // 존재하는 보스의 개수
+    public int boss_num = 0; // 진행중인 보스의 인덱스
+    public const int boss_cnt = 5; // 존재하는 보스의 개수
     
-    // 보스 정보. 이름, 속성으로 이루어짐.
-    public string[,] boss_info = new string[,] { { "Dog" , "1" }, { "ForgetMeNot", "3" }, { "PartTime", "2" } }; 
-
-    public bool[] boss_clear_info = new bool[boss_cnt]; // 보스 클리어 여부 확인
+    // 보스 정보. 이름으로 이루어짐.
+    public string[] boss_info = new string[] { "Dog", "PartTime", "ForgetMeNot", "Criminal", "Hoa"}; 
     
     public bool is_back = false; //클리어 후 돌아오는 길인지 corrider에서 확인하기 위한 변수
 
@@ -32,8 +30,6 @@ public class GameManager : MonoBehaviour
         {
             Destroy(gameObject);
         }
-
-
         DontDestroyOnLoad(gameObject);
     }
 }
