+++ EPOH/Assets/Scripts/NotUsing.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de4f9fee2257a58448b28ea720d9e6dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
