+++ EPOH/Assets/Scripts/TutorialEvent/Event1Tutorial.cs
@@ -6,45 +6,32 @@ using TMPro;
 
 public class Event1Tutorial : MonoBehaviour
 {
-    public TalkAction action;
-    public PlayerController player_controller; 
-
-    public GameObject tutorial_panel; 
-    //public TMP_Text tutorial_text;
-
-    private bool isTutorialShown = false;
-
-
+    private TalkAction action;
+    private PlayerController player_controller;
+    public GameObject tutorial_panel;
 
     private bool space_pressed = false;
+
     
     private void Start()
     {
         tutorial_panel.SetActive(false);
+        player_controller = GameObject.FindGameObjectWithTag("Player").GetComponent<PlayerController>();
         action = GameObject.FindGameObjectWithTag("TalkManager").GetComponent<TalkAction>();
     }
 
     private void Update()
     {
-        // GameManager.instance.story_info가 1이면 tutorial_panel 활성화
-        if (GameManager.instance.story_info == 1 && !isTutorialShown)
+        // GameManager의 story_info가 1이고 (스토리 1이 진행되었고) tutorial_info가 0이면 tutorial_panel 활성화
+        if (GameManager.instance.story_info == 1 && GameManager.instance.tutorial_info == 0)
         {
-            if(tutorial_panel.activeSelf == false && action.is_talking == false)
-            {
-               StartCoroutine(showTutorial());
-            }
-
-            
-            if (Input.GetKeyDown(KeyCode.Space) && tutorial_panel.activeSelf) // 스페이스 키를 눌렀는지 체크
-            {
-                space_pressed = true;
-
-            }
-            
+            StartCoroutine(showTutorial());
+            GameManager.instance.tutorial_info++; // tutorial 카운트를 올림
         }
-        else
+
+        if (Input.GetKeyDown(KeyCode.Space) && tutorial_panel.activeSelf) // 스페이스 키를 눌렀는지 체크
         {
-            space_pressed = false;  // 스페이스 키를 눌렀다는 체크를 초기화
+            space_pressed = true;
         }
     }
 
@@ -53,14 +40,9 @@ public class Event1Tutorial : MonoBehaviour
         tutorial_panel.SetActive(true); 
         player_controller.is_interacting = true; 
 
-        //tutorial_text.text = "튜토리얼- ";
-        //yield return new WaitForSeconds(0.6f); 
-        //StartCoroutine(printTutorialTextEffect("좌우 방향키로 이동할 수 있습니다. Z키로 점프할 수 있습니다. 점프는 최대 2번까지 연속으로 가능합니다.", 0.05f));
         yield return StartCoroutine(waitForKeyPress());
         tutorial_panel.SetActive(false); 
         player_controller.is_interacting = false;
-
-        isTutorialShown = true; // 튜토리얼이 보여졌음을 표시
     }
 
     IEnumerator waitForKeyPress() // Space 키를 누르면 다음 대사로 넘어가는 함수
@@ -71,15 +53,5 @@ public class Event1Tutorial : MonoBehaviour
         }
         space_pressed = false; // Space 키를 눌렀다는 체크를 초기화
     }
-
-    /*IEnumerator printTutorialTextEffect(string text, float delay)
-    {
-        for (int i = 0; i < text.Length; i++)
-        {
-            tutorial_text.text += text[i];
-            yield return new WaitForSeconds(delay);
-        }
-    }
-    */
 
 }
